user specific handling

+++ Application/Interfaces/IBookRepository.cs
@@ -4,10 +4,10 @@ namespace Application.Interfaces;
 
 public interface IBookRepository
 {
-    Task<Book?> GetBookAsync(ulong id);
-    Task<ulong> AddBookAsync(Book book);
+    Task<Book?> GetBookAsync(int id);
+    Task<int> AddBookAsync(Book book);
     Task UpdateBookAsync(Book book);
-    Task DeleteBookAsync(ulong id);
+    Task DeleteBookAsync(int id);
     Task<List<Book>> GetAllBookAsync();
-    Task<List<Book>> GetBooksByUserIdAsync(ulong id);
+    Task<List<Book>> GetBooksByUserIdAsync(int id);
 }
