+++ Application/Interfaces/IUserRepository.cs
@@ -4,6 +4,6 @@ namespace Application.Interfaces;
 
 public interface IUserRepository
 {
-    Task<ulong> AddUserAsync(User user);
+    Task<int> AddUserAsync(User user);
     Task<User?> GetUserByEmailAsync(string email);
 }
